requirements.txt
Django==2.2.12
django-environ==0.4.5
djangorestframework==3.11.0
notebook==6.0.2
numpy==1.18.4
pandas==0.25.3
pomegranate==0.11.2
requests==2.23.0
